requirements.txt
Django>=2.1
django-filter>=2.0.0
djangorestframework>=3.8.2
drf-nested-routers>=0.91
coreapi>=2.3.3
coreapi-cli>=1.0.9
markdown>=3.0.1
